+++ src/main/java/domain/PlayingCardValue.java
@@ -1,4 +1,4 @@
-package domain.card;
+package domain;
 
 public enum PlayingCardValue {
     SMALL_ACE(1), TWO(2), THREE(3), FOUR(4), FIVE(5), SIX(6), SEVEN(7), EIGHT(8), NINE(9), TEN(10), JACK(10), QUEEN(10), KING(10), DEFAULT_ACE(11);
