+++ src/main/java/domain/constant/GameOption.java
@@ -1,4 +1,4 @@
-package domain.card.constant;
+package domain.constant;
 
 public class GameOption {
     public static final int BLACKJACK_CONDITION = 21;
